Use derive-contract in admin impl
