--- lib/composable-admin/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "composable-admin"
-version = "0.1.0"
-authors = ["Asparuh Kamenov <asparuh@hack.bg>"]
-edition = "2018"
-
-[[example]]
-name = "contract"
-crate-type = ["staticlib"]
-test = true
-
-[dependencies]
-fadroma = { path = "../", features = [ "scrt-storage" ] }
-require-admin = { path = "../require-admin", optional = true }
-
-schemars = "0.7"
-serde = { version = "1.0.103", default-features = false, features = ["derive"] }
-
-[features]
-default = ["derive"]
-derive = ["require-admin"]
