--- lib/composable-admin/src/lib.rs
@@ -1,5 +1,0 @@
-#[cfg(feature = "derive")]
-pub use require_admin::require_admin;
-
-pub mod multi_admin;
-pub mod admin;
