+++ lib/composable_admin.rs
@@ -1,0 +1,4 @@
+#[path ="./composable-admin/admin.rs"]
+pub mod admin;
+#[path ="./composable-admin/multi_admin.rs"]
+pub mod multi_admin;
